divide profile.tsx
