+++ src/Pages/Profile/ProfileComponents/TwinklingStars.tsx
@@ -1,0 +1,39 @@
+import React from 'react';
+
+interface TwinklingStarsProps {
+  count?: number;
+  className?: string;
+}
+
+export const TwinklingStars: React.FC<TwinklingStarsProps> = ({ 
+  count = 30, 
+  className = '' 
+}) => {
+  const stars = Array.from({ length: count }, (_, i) => ({
+    id: i,
+    size: Math.random() * 3 + 1,
+    x: Math.random() * 100,
+    y: Math.random() * 100,
+    delay: Math.random() * 5,
+    duration: Math.random() * 4 + 2,
+  }));
+
+  return (
+    <div className={`absolute inset-0 pointer-events-none overflow-hidden ${className}`}>
+      {stars.map((star) => (
+        <div
+          key={star.id}
+          className="absolute rounded-full bg-gradient-to-r from-[#f1b3be] to-[#ffe0db] animate-twinkle opacity-60"
+          style={{
+            width: `${star.size}px`,
+            height: `${star.size}px`,
+            left: `${star.x}%`,
+            top: `${star.y}%`,
+            animationDelay: `${star.delay}s`,
+            animationDuration: `${star.duration}s`,
+          }}
+        />
+      ))}
+    </div>
+  );
+};
