+++ src/Pages/Profile/ProfileComponents/index.ts
@@ -1,0 +1,17 @@
+// Exportar todos los tipos
+export * from './types';
+
+// Exportar utilidades
+export { apiClient } from './apiClient';
+
+// Exportar componentes
+export { TwinklingStars } from './TwinklingStars';
+export { LoadingScreen } from './LoadingScreen';
+export { ProfileBreadcrumbs } from './ProfileBreadcrumbs';
+export { ProfileHeader } from './ProfileHeader';
+export { ProfileAvatar } from './ProfileAvatar';
+export { Notification } from './Notification';
+export { DeleteConfirmModal } from './DeleteConfirmModal';
+export { ProfileTab } from './ProfileTab';
+export { PrivacyTab } from './PrivacyTab';
+export { SecurityTab } from './SecurityTab';
